Move data and layout files out of the input dir
This allows later converting the input dir into a git submodule. This
way the site's source code can be public while keeping the posts in a
private git repo. (And this is nice because this way the private repo
can contain draft posts and posts with a future publication date.)